requirements.txt
openai>=1.35.0
pydantic>=2.7.0
python-dotenv>=1.0.1
typer>=0.12.3
orjson>=3.10.0
yfinance>=0.2.40
httpx>=0.27.0
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
playwright>=1.40.0
